gro: Fix ABI change for CVE-2016-7039 fix
The fix for CVE-2016-7039 (unbounded recursion in GRO) added a 4-bit
bitfield in struct napi_gro_cb where there used to be padding, so
the structure layout hasn't changed.

In case of a version mismatch between kernel and modules, two things
could go wrong, but they are not disastrous:

- the recursion count is not incremented and checked when it should
  be, so the fix is incomplete - no worse than before
- the recursion count is not reset, so GRO is not effective - only
  a performance problem

So hide this change from genksyms.


Gbp-Pq: Topic debian
Gbp-Pq: Name gro-fix-abi-change-for-cve-2016-7039-fix.patch