drm/i915/cmdparser: Limit clflush to active cachelines
commit 504ae4024131c5a01c3ce8382d49b801375e039c upstream.

We only need to clflush those cachelines that we have validated to be
read by the GPU. Userspace typically fills the batch length in
correctly, the exceptions tend to be explicit tests within igt.

v2: Use ptr_mask_bits() to make Mika happy
v3: cmd is not advanced on MI_BBE, so make sure to include an extra
dword in the clflush.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/20170310115518.13832-1-chris@chris-wilson.co.uk
Signed-off-by: Jon Bloomfield <jon.bloomfield@intel.com>

Gbp-Pq: Topic bugfix/x86/i915
Gbp-Pq: Name 0006-drm-i915-cmdparser-Limit-clflush-to-active-cacheline.patch