mac80211: Avoid ABI change in 4.9.53
Commit 531682159092 ("mac80211: fix VLAN handling with TXQs") replaced
a union with two separate members in ieee80211_tx_info::control.  The
one which changed offset (enqueue_time) doesn't appear to be accessed
outside of the mac80211 module, so this shouldn't be an ABI change for
drivers.  Therefore hide the ABI change from genksyms.


Gbp-Pq: Topic debian
Gbp-Pq: Name mac80211-avoid-abi-change-in-4.9.53.patch