s390/mm: Avoid ABI change in 4.9.52
Commit 60f07c8ec5fa "s390/mm: fix race on mm->context.flush_mm" added
a field to the s390 definition of mm_context_t.  This is embedded in
the arch-indepdendent struct mm_struct, so adding a new field results
in an incompatible change to the layout of the larger struct.

Move the new field out of mm_context_t to the end of mm_struct, and
hide the change from genksyms.


Gbp-Pq: Topic debian
Gbp-Pq: Name s390-mm-avoid-abi-change-in-4.9.52.patch