mmc: sdio: Avoid ABI change in 4.9.54
Commit 5ef1ecf060f2 "mmc: sdio: fix alignment issue in struct
sdio_func" changed the type of sdio_func::tmpbuf from array to
pointer.  On 64-bit architectures this also changes the size of the
field.  It appears that tmpbuf is only used by the MMC core, and the
structure is always allocated there.  So put padding in the old
position, move the pointer to the end of the structure, and hide this
from genksyms.

Just in case I missed a user elsewhere, which might end up accessing
memory off the end of an old structure, rename the field to
core_tmpbuf (and update its users in the core).


Gbp-Pq: Topic debian
Gbp-Pq: Name mmc-sdio-avoid-abi-change-in-4.9.54.patch