x86/cpu: Avoid ABI change in 4.9.83
The type of cpuinfo_x86::x86_cache_size was changed from int to
unsigned int, and the "unknown" value was changed from -1 to 0.  Only
one in-tree user of the field seems to have checked for the "unknown"
value, so this is probably not an ABI change in practice.  Just hide
the type change from genksyms.


Gbp-Pq: Topic debian
Gbp-Pq: Name x86-cpu-avoid-abi-change-in-4.9.83.patch