ext4: add corruption check in ext4_xattr_set_entry()
In theory this should have been caught earlier when the xattr list was
verified, but in case it got missed, it's simple enough to add check
to make sure we don't overrun the xattr buffer.

This addresses CVE-2018-10879.

https://bugzilla.kernel.org/show_bug.cgi?id=200001

Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Reviewed-by: Andreas Dilger <adilger@dilger.ca>
[bwh: Backported to 4.9:
 - Add inode parameter to ext4_xattr_set_entry()
 - Return on error; there is no 'out' label and no cleanup is needed
 - Adjust context]

Gbp-Pq: Topic bugfix/all
Gbp-Pq: Name ext4-add-corruption-check-in-ext4_xattr_set_entry.patch