x86/cpu: Avoid ABI change in 4.9.125
Commit ef3d45c95764 "x86/speculation/l1tf: Increase l1tf memory limit
for Nehalem+" added a new field to the end of cpuinfo_x86.  This
structure is always allocated statically in the core kernel, and not
embedded in anything larger, so extending it is safe.

Modules won't use the new field; make sure of that by guarding it
with #ifndef MODULE as well as hiding it from genksyms.


Gbp-Pq: Topic debian
Gbp-Pq: Name x86-cpu-avoid-abi-change-in-4.9.125.patch