mm: Avoid ABI change in 4.9.128
Commit 84580567f1f8 "mm: get rid of vmacache_flush_all() entirely"
changed the type (and size) of mm_struct::vmacache_seqnum and
task_struct::vmacache_seqnum.

This field is only used by the built-in vmacache code, and the
structures are always allocated by the core kernel, so it should
be safe to move the fields to the end of each structure and add
padding in place of the old location.

However, task_struct is really dynamically-sized on s390 and x86
and we can't safely access another field at the end.  Instead,
define a new struct task_struct_ext containing the larger field,
and:

- Add space for this to the end of struct task_struct.  s390 and
  x86 calculate the dynamic structure size by subtracting from
  sizeof(struct task_struct) rather than using offsetof(), so they
  will include this extra space in the dynamic size.
- Define a task_ext() function to find the task_struct_ext at the
  end of a task_struct, using the dynamic structure size.
- Use task_ext() in the vmacache code to access the larger field.
- [x86] Adjust the structure layout check in
  fpu__init_task_struct_size for this.


Gbp-Pq: Topic debian
Gbp-Pq: Name mm-avoid-abi-change-in-4.9.128.patch