x86/fpu: Avoid ABI change in 4.9.133
Commit 60696d91bbd6 "x86/fpu: Remove struct fpu::counter" did just
what it said.  This is not a real ABI change as that member was
private and is a single byte that becomes a padding byte.  However
genksyms thinks otherwise - and it's just possible that later changes
might cause a real size change.  So add the member back, but mark it
clearly as padding.


Gbp-Pq: Topic debian
Gbp-Pq: Name x86-fpu-avoid-abi-change-in-4.9.133.patch