reset: Avoid ABI changes in 4.9.144
Commit ca58e3b2fc54 "reset: make optional functions really optional"
added an extra parameter to the __of_reset_control_get() and
__devm_reset_control_get() functions.  Rename these functions with a
suffix "_5" (for 5 parameters) and add back the old functions as
wrappers around them.  Use macros to ensure the new functions are used
by newly built code.

Commit f10a9deb22ee "reset: make device_reset_optional() really
optional" changed device_reset() from an extern exported function
(when CONFIG_RESET_CONTROLLER is enabled) to a static inline wrapper.
Keep it as a wrapper but change it back to an extern exported
function.


Gbp-Pq: Topic debian
Gbp-Pq: Name reset-avoid-abi-changes-in-4.9.144.patch