ARM: Avoid ABI change in 4.9.139
Commit 901e325f772f "ARM: bugs: add support for per-processor bug
checking" added a new member to struct processor.  This structure
is always instantiated in built-in code and the new member is only
used in built-in code.  Therefore we can safely add the new member
at the end instead.

Move it to the end and hide it from genksyms.  Also hide it when
building modules, to make sure they really don't use it.


Gbp-Pq: Topic debian
Gbp-Pq: Name arm-avoid-abi-change-in-4.9.139.patch